Renamed occupation_helpers
